Uses correct option names on installation.
The option names were different on installation than those used through the
plugin. This causes the default values to not appear in places like the
configuration form or the public theme header.

I opted to change the names in the installation hook instead of changing them
elsewhere and adding a migration, since the plugin has yet to be tagged and
released. Users will need to uninstall and reinstall the plugin to get the
correct options, however.